Internal reorganization of files, adding a "common utils" file
